[Refactoring ⚙️] PlaceMap으로 카카오맵 컴포넌트 따로 분리
